major: remove need of db, make blocking happen on client-side
This is a major commit, since it removes the need for a database.
Before, the database was used to register users so other users could not fake the username/identifier.
But I got the conclusion it isn't needed. So now anyone can use any identifier they want,
as long as its not occupied. The blocking of users now happens on the receiver side, where the list of the blocked users
is now in a new file called `.blocks`.